Time: 11 ms (42.22%), Space: 22.9 MB (73.86%) - LeetHub

+++ 0907-koko-eating-bananas/0907-koko-eating-bananas.cpp
@@ -21,7 +21,7 @@ public:
         int ans = 0;
         while(low <= high){
             int mid = low + (high - low) / 2;
-            int totalh = totalHours(piles,mid);
+            long long totalh = totalHours(piles,mid);
      
             if(totalh <= h){
                 ans = mid;
